requirements.txt
anthropic>=0.25.0
databricks-sdk>=0.20.0
databricks-sql-connector>=3.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.7.0
typer>=0.9.0
pandas>=2.1.0
pytest>=7.4.0
PyYAML>=6.0.1
httpx>=0.25.0
